Easier border dragging and more customizability
Added the ability to perform a long press on the reachability space and
return to app in the view(thanks to LastApp tweak). User can now choose
a color in regards to where the drag zone is. Goal is to get the user
accustom to where the drag zone is. Can easily disable from the tweak
preferences once not needed.